requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
kombu==5.2.4
packaging==23.1
pluggy==1.0.0
prompt-toolkit==3.0.38
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
redis==4.5.5
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.2
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
